# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Documentation
  description: Документация API для всех операций, включая аутентификацию, управление картами, блокировку карт и управление пользователями
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер для разработки
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponseDto:
      type: object
      description: Объект ответа, содержащий информацию о пользователе
      required:
        - id
        - login
        - roleName
        - isEnabled
        - isNonLocked
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 10
        login:
          type: string
          description: Логин пользователя
          example: user123
        roleName:
          type: string
          description: Название роли пользователя
          example: ADMIN
        isEnabled:
          type: boolean
          description: Флаг, указывающий, активен ли пользователь
          example: true
        isNonLocked:
          type: boolean
          description: Флаг, указывающий, не заблокирован ли пользователь
          example: true
        cards:
          type: array
          description: Список карт пользователя (может отсутствовать в некоторых ответах)
          items:
            $ref: '#/components/schemas/CardResponseDto'
...
    CardResponseDto:
      type: object
      description: Объект ответа, содержащий информацию о карте
      required:
        - id
        - maskedCardNumber
        - ownerLogin
        - expiryDate
        - statusName
        - balance
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор карты
          example: 10
        maskedCardNumber:
          type: string
          description: Маскированный номер карты, отображающий только последние 4 цифры
          example: "************1234"
        ownerLogin:
          type: string
          description: Логин владельца карты
          example: user123
        expiryDate:
          type: string
          pattern: ^\d{4}-\d{2}$
          description: Дата истечения срока действия карты в формате yyyy-MM
          example: 2025-08
        statusName:
          type: string
          description: Название статуса карты
          example: ACTIVE
        balance:
          type: number
          format: double
          description: Текущий баланс на карте
          example: 1500.75
...
    CardStatusCode:
      type: string
      description: Код статуса карты
      enum:
        - ACTIVE
        - INACTIVE
        - BLOCKED
...
    CardBlockResponseDto:
      type: object
      description: Объект ответа для операции блокировки карты
      required:
        - id
        - requestDate
        - cardId
        - userId
        - status
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор запроса на блокировку
          example: 1
        requestDate:
          type: string
          format: date-time
          description: Дата и время создания запроса на блокировку
          example: 2025-08-13T20:37:00
        cardId:
          type: integer
          format: int64
          description: Идентификатор карты, подлежащей блокировке
          example: 10
        userId:
          type: integer
          format: int64
          description: Идентификатор пользователя, связанного с картой
          example: 100
        status:
          $ref: '#/components/schemas/RequestStatusCode'
        adminComment:
          type: string
          description: Комментарий администратора к запросу на блокировку
          example: Карта заблокирована по запросу пользователя
    RequestStatusCode:
      type: string
      description: Статус запроса на блокировку
      enum:
        - PENDING
        - APPROVED
        - REJECTED
...
    Page:
      type: object
      description: Объект пагинации для списков
      properties:
        content:
          type: array
          description: Список элементов на странице
          items:
            oneOf:
              - $ref: '#/components/schemas/CardResponseDto'
              - $ref: '#/components/schemas/CardBlockResponseDto'
              - $ref: '#/components/schemas/UserResponseDto'
        page:
          type: integer
          description: Номер текущей страницы
          example: 0
        size:
          type: integer
          description: Размер страницы
          example: 10
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
          example: 100
        totalPages:
          type: integer
          description: Общее количество страниц
          example: 10
paths:
...
  /cards/{id}:
    get:
      summary: Получение карты по ID
      description: Возвращает информацию о карте по её идентификатору
      tags:
        - Управление картами
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор карты
          example: 10
      responses:
        '200':
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '400':
          description: Некорректный запрос
          content: {}
        '401':
          description: Неавторизован
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Карта не найдена
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
    delete:
      summary: Удаление карты
      description: Удаляет карту по её идентификатору
      tags:
        - Управление картами
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор карты
          example: 10
      responses:
        '204':
          description: Карта удалена
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Карта не найдена
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /cards/{id}/my:
    get:
      summary: Получение карты текущего пользователя
      description: Возвращает информацию о карте текущего пользователя по её идентификатору
      tags:
        - Управление картами
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор карты
          example: 10
      responses:
        '200':
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponseDto'
        '400':
          description: Некорректный запрос
          content: {}
        '401':
          description: Неавторизован
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Карта не найдена
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /cards/my:
    get:
      summary: Получение списка карт текущего пользователя
      description: Возвращает список карт текущего пользователя с возможностью фильтрации и пагинации
      tags:
        - Управление картами
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы
          example: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Размер страницы
          example: 10
        - name: lastFourDigits
          in: query
          required: false
          schema:
            type: string
          description: Последние четыре цифры номера карты
          example: 1234
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatusCode'
          description: Статус карты
          example: ACTIVE
        - name: minBalance
          in: query
          required: false
          schema:
            type: number
            format: double
          description: Минимальный баланс на карте
          example: 100.00
        - name: maxBalance
          in: query
          required: false
          schema:
            type: number
            format: double
          description: Максимальный баланс на карте
          example: 1000.00
      responses:
        '200':
          description: Список карт найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          description: Некорректный запрос
          content: {}
        '401':
          description: Неавторизован
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
...
  /card-blocks/{id}:
    get:
      summary: Получение запроса на блокировку по ID
      description: Возвращает информацию о запросе на блокировку карты по его идентификатору
      tags:
        - Блокировка карт
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор запроса на блокировку
          example: 1
      responses:
        '200':
          description: Успешное получение запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockResponseDto'
        '400':
          description: Некорректный запрос
          content: {}
        '401':
          description: Ошибка аутентификации
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Запрос не найден
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
    delete:
      summary: Удаление запроса на блокировку
      description: Удаляет запрос на блокировку карты по его идентификатору
      tags:
        - Блокировка карт
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор запроса на блокировку
          example: 1
      responses:
        '204':
          description: Запрос успешно удален
          content: {}
        '400':
          description: Некорректный запрос
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Запрос не найден
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /card-blocks/{id}/my:
    get:
      summary: Получение запроса на блокировку текущего пользователя
      description: Возвращает информацию о запросе на блокировку карты, связанного с текущим пользователем
      tags:
        - Блокировка карт
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор запроса на блокировку
          example: 1
      responses:
        '200':
          description: Успешное получение запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockResponseDto'
        '400':
          description: Некорректный запрос
          content: {}
        '401':
          description: Недействительный или отсутствующий токен
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Запрос не найден
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /card-blocks/my:
    get:
      summary: Получение списка запросов на блокировку текущего пользователя
      description: Возвращает страницу запросов на блокировку карт, связанных с текущим пользователем
      tags:
        - Блокировка карт
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Номер страницы (начинается с 0)
          example: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Размер страницы
          example: 10
      responses:
        '200':
          description: Успешное получение списка запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          description: Некорректный запрос
          content: {}
        '401':
          description: Недействительный или отсутствующий токен
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
...
  /card-blocks/{id}/approve:
    patch:
      summary: Одобрение запроса на блокировку
      description: Одобряет запрос на блокировку карты с указанным комментарием
      tags:
        - Блокировка карт
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор запроса на блокировку
          example: 1
        - name: comment
          in: query
          required: true
          schema:
            type: string
          description: Комментарий администратора
          example: Одобрено администратором
      responses:
        '200':
          description: Запрос успешно одобрен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockResponseDto'
        '400':
          description: Некорректный запрос
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Запрос не найден
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
  /card-blocks/{id}/reject:
    patch:
      summary: Отклонение запроса на блокировку
      description: Отклоняет запрос на блокировку карты с указанным комментарием
      tags:
        - Блокировка карт
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор запроса на блокировку
          example: 1
        - name: comment
          in: query
          required: true
          schema:
            type: string
          description: Комментарий администратора
          example: Отклонено из-за недостаточной информации
      responses:
        '200':
          description: Запрос успешно отклонен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockResponseDto'
        '400':
          description: Некорректный запрос
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Запрос не найден
          content: {}
        '500':
          description: Внутренняя ошибка сервера
          content: {}
...